chore: reexport `CrateName` through `nargo` (#6177)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

Ideally `nargo_cli` would not need to directly hook into
`noirc_frontend` but a number of imports it needs are not reexported by
the packages which are supposed to be the entrypoints to the noir
compiler (`noirc_driver` and `nargo`). This PR tackles a couple of cases
of these issues.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.