feat(perf): Simplify the cfg after DIE (#6184)
# Description

## Problem\*

Part of general effort to reduce Brillig bytecode sizes.

## Summary\*

While working other SSA opts, I noticed in a couple tests we had
simplified blocks down to only their terminator, but those empty blocks
still existed in our final SSA. It looks to be due to previous passes,
and specifically DIE simplifying down to an empty block which was not
caught in the initial CFG simplification.

I pushed this PR wanted to see how much of a benefit simplifying at the
end would be across our tests. It is a whole extra SSA pass, however, of
all our passes, simplify CFG shouldn't be very heavy and wouldn't be too
bad to run again if it provides decent benefits.

## Additional Context



## Documentation\*

Check one:
- [X] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.