feat: Inclusive for loop (#6200)
# Description

## Problem\*

Resolves #6171

## Summary\*

Adds support for inclusive for syntax, e.g. `for i in 0..=10`.

## Additional Context

- Added a `ForBounds` struct to handle the desugaring of `start..=end`
into `start..end+1`.
- `ForBounds` has an `inclusive` flag; I thought about adding a
`ForRange::RangeInclusive` similar to `std::ops::RangeInclusive` in
Rust, but then I thought `for` loops will not have the other variants
like `..end`, `start..` and `..`, which wouldn't have two separate
expressions, and the flag is simpler.
- Transforming an inclusive range to Hir and back makes it
non-inclusive, but I think that's fine because the block created by
`ForRange::into_for` to iterate over the array is not turned back
either.


## Documentation\*

Check one:
- [ ] No documentation needed.
- [x] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.