feat: visibility for impl functions (#6179)
# Description

## Problem

Part of #4515

## Summary

~After searching for a while I found that to make this work we only
needed to change one line~

## Additional Context

I didn't include tests for now, but I'll add them later because we first
need to figure out which `impl` functions should be marked as `pub` or
`pub(crate)` in the stdlib. I'm not familiar with the stdlib though I
probably can't do it. That said, if you open the standard library with a
nargo compiled from this PR you already get some warnings, so those
functions should at least not be private.


## Documentation

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Tom French <tom@tomfren.ch>