fix: don't warn twice when referring to private item (#6216)
# Description

## Problem

We were producing "this item is private" errors twice.

## Summary

The problem was that the code looked up a path twice, and errored each
time it found an error in that lookup.

Now the code still looks up the path twice, but only errors once (if the
first lookup succeeds, we report the optional error of that lookup,
otherwise we do the same with the second one).

## Additional Context

Also applied some refactors (made a tuple return value its own struct,
and made an enum variant use a struct instead of a tuple too).

This fixes the issue but we are still doing two Path lookups here, which
means there's an optimization opportunity to only do the lookup once.
But this can be done in a separate PR.

## Documentation

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.