fix: add missing visibility for auto-import names (#6205)
# Description

## Problem

Even though we started tracking visibility for many items, this
visibility wasn't registered for auto-import.

## Summary

## Additional Context

I was thinking that maybe registering things for auto-import isn't
necessary. We could traverse all types and all functions in all modules
and check if they match... but maybe that'll be much slower (right now
we group all items by name so maybe iterating that is faster). In any
case we could explore that other option later on.

## Documentation

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.