fix(frontend): Do not warn when a nested struct is provided as input to main (#6239)
# Description

## Problem\*

No issue as found while working on other tasks.

Compile `nested_array_dynamic` and you will see a warning the `Bar` is
never constructed. However, `Bar` is an internal field to the `Foo`
struct which is provided as inputs and is thus constructed.

## Summary\*

When marking a parameter type as used we also go over all of the struct
fields to account for any nested structs.

## Additional Context



## Documentation\*

Check one:
- [X] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.