feat: allow `unconstrained` after visibility (#6246)
# Description

## Problem

We currently parse `unconstrained pub fn ...` but ideally we'd want it
to be `pub unconstrained fn ...`

See this comment too:
https://github.com/noir-lang/noir/pull/6180#discussion_r1790221742

## Summary

This PR implements the above, but actually all of these:
- We now allow `unconstrained pub fn ...` and `pub unconstrained fn`
(the former for backwards compatibility)
- The formatter will change `unconstrained pub fn` to `pub unconstrained
fn` (so after some time we can remove the backwards-compatibility case)
- The formatter will now format a function "header" (I didn't know what
to call it): that's the fn outer comments, attributes, and modifiers.
For example before this PR if you wrote `pub fn foo() {}` it was left
untouched, but now it's changed to `pub fn foo() {}`.

## Additional Context

None.

## Documentation

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.