chore: add regression test for #5756 (#5770)
# Description

## Problem\*

Adds regression tests for #5756 

## Summary\*

This PR is some investigation for #5756 as I'm not sure on the
underlying cause as the keccakf1600 instructions _should_
be deduplicated. I'm getting different behaviour in the constant folding
pass test compared to `test_programs` which I'm a little confused by.

Posting this so other people can investigate as well.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.