feat: don't crash LSP when there are errors resolving the workspace (#6257)
# Description

## Problem

Before this PR, LSP would crash for assumed workspaces (when we can't
find a Nargo.toml) or when there were errors in the Nargo.toml.

## Summary

Now LSP doesn't crash in the above cases:
- for assumed workspaces we don't crash (we used to not crash but I made
some changes in the past that broke this)
- for errors in Nargo.toml we now output to STDERR, so you can see the
error in the output. This is what Rust Analyzer does, except that they
also have a "rust-analyzer" thing at the bottom that becomes yellow when
there's an error... I didn't implement this (it probably requires some
code on the extension side) but I think if something doesn't work you'd
check the logs or ask, and it's worse if it crashes.

## Additional Context

## Documentation

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.