fix!: Integer division is not the inverse of integer multiplication (#6243)
# Description

## Problem\*

Resolves https://github.com/noir-lang/noir/issues/6242

## Summary\*

Making this PR for visibility to show that fixing this issue breaks our
serialize test.

Notably this fix allows rounding arithmetic generics using the `/ N * N`
pattern which was simplified away previously.

## Additional Context

I initially removed the cancellation of / and * entirely but found that
this breaks even more tests. So I added `approx_inverse` in a few cases
that only involve constants that I thought may be correct but we still
fail `serialize` even with these cases.

## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.