chore(refactor): Remove optional global_id from methods that can get it from the definition (#6283)
# Description

## Problem\*

Followup to https://github.com/noir-lang/noir/pull/6282

## Summary\*
Some methods in the `Elaborator` received a `global_id:
Option<GlobalId>` and a `definition: DefinitionKind` parameter. We found
that `Elaborator::elaborate_let` actually created `definition` to carry
the same global ID as it was given, so `global_id: Some(id)` and
`definition: DefinitionKind::Global(id)` always carried the same value.
We added an
[assertion](https://github.com/noir-lang/noir/pull/6282/commits/1d2bd36ab5fb38ebab28b8262dbf13f8020ee461)
that checked that they are never different.

The next logical step is to remove the possibility of confusion by only
carrying the ID in the definition.

## Additional Context

## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [ ] I have tested the changes locally.
- [ ] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.