fix: enforce correctness of decompositions performed at compile time (#6278)
# Description

## Problem\*

Resolves https://github.com/noir-lang/noir/issues/6244

## Summary\*

We were not checking that a radix decomposition simplification performed
at compile time was valid, resulting in a truncation to fit the result
array. This PR prevents the simplification in this case so we can
instead error at runtime.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.