feat(ssa): Various mem2reg reverts to reduce memory and compilation time (#6307)
# Description

## Problem\*

Resolves #6267

Pushing as a draft to see bytecode size regressions we get.

## Summary\*

This version for the code in #6267 took less than 2 gigs of ram and
4.66s to compile. The code on master takes ~20 gigs and 94.3s to
compile. The noir-contracts workspace with this mem2reg takes ~20 gigs
of ram and 94.3s to compile vs. >70 gigs of ram and 4.5 min to compile
w/ the mem2reg on master.

## Additional Context



## Documentation\*

Check one:
- [X] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Tom French <tom@tomfren.ch>